+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,6 +1,7 @@
 package nextstep.jwp.controller;
 
 import nextstep.jwp.service.UserService;
+import org.apache.catalina.servlet.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.StatusLine;
