+++ tomcat/src/main/java/org/apache/catalina/servlet/Controller.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.controller;
+package org.apache.catalina.servlet;
 
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
